update manage task
