+++ resources/views/dashboard/video/fixed-video.blade.php
@@ -100,6 +100,42 @@
                         </form>
                     </div>
                 </div>
+                <div class="copy hidden" id="copy">
+                    <h5 class="mb-0 text-[#009FB2] text-lg font-semibold">Choose folder to move selected to</h5>
+                    <div class="mt-3" list-folder>
+                        <div class="items-center w-full">
+                            <div
+                                class="relative flex flex-wrap mr-3 items-stretch transition-all rounded-lg ease">
+                                <span
+                                    class="text-sm ease leading-5.6 absolute z-10 -ml-px flex h-full items-center whitespace-nowrap rounded-lg rounded-tr-none rounded-br-none border border-r-0 border-transparent bg-transparent py-2 px-2.5 text-center font-normal text-slate-500 transition-all">
+                                  <i class="fa fa-search"></i>
+                                </span>
+                                <input type="text" search-folder
+                                       class="pl-9 text-sm focus:shadow-primary-outline ease leading-5.6 relative -ml-px block min-w-0 flex-auto rounded-lg bg-[#142132] text-white bg-clip-padding py-2 pr-3 transition-all placeholder:text-gray-500 focus:border-blue-500 focus:outline-none focus:transition-shadow"
+                                       placeholder="Search folder"/>
+                            </div>
+                        </div>
+                        <div id="list-move-folder"
+                             class="grid grid-cols-3 sm:grid-cols-4 lg:grid-cols-6 gap-4 text-white mt-2 min-h-80 max-h-80 overflow-auto">
+                            @foreach($folders as $folder)
+                                <div  folder data-folder-id="{{$folder -> id}}" class="item-folder text-center cursor-pointer hover:text-transparent hover:bg-gradient-to-r bg-clip-text  from-[#009fb2] to-[#009fb2]">
+                                    <div class="text-center">
+                                        <i class="material-symbols-outlined text-3xl">folder</i>
+                                    </div>
+                                    <h5 name-folder>
+                                        {{$folder -> name_folder}}
+                                    </h5>
+                                </div>
+                            @endforeach
+                        </div>
+                        <form class="pt-2 w-max text-white">
+                            <button move-to-folder type="submit" disabled
+                                    class="bg-[#142132] rounded-lg w-max font-semibold text-md py-2.5 px-5">
+                                Move To Folder
+                            </button>
+                        </form>
+                    </div>
+                </div>
             </div>
         </div>
     </div>
